Refactor: 필요없는 코드 제거 - 중복호출
